requirements.txt
selenium
requests
python-dotenv
pandas
openpyxl
beautifulsoup4
webdriver-manager
openai
ipython
shiny
shinyswatch
